Moved responder component to it's own file
